Sign out and remember the user implementation
when the user closes the application , we create a property file and save the current user data. when the user try to open the application again we read this info, perform sign in for him and switch to he chat scene,

we also handled the sign out action
when the user sign out  we change the data in the property file to force the user to log in next time he uses the application.

note:
modification to be added tomorrow inshallah
1- encrypt the data
2- change the location of calling the handle log out method.